pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Physical AI Textbook Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.2,<0.125.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "qdrant-client",
    "sentence-transformers",
    "google-cloud-translate",
    "psycopg2-binary",
    "python-dotenv",
    "openai",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "sqlalchemy",
    "alembic",
    "pydantic",
    "pydantic-settings",
]

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rofan-Jlove <kannethchamp@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.124.2,<0.125.0"
uvicorn = {extras = ["standard"], version = ">=0.38.0,<0.39.0"}
langchain = "*"
langchain-community = "*"
langchain-openai = "*"
qdrant-client = "*"
sentence-transformers = "*"
google-cloud-translate = "*"
psycopg2-binary = "*"
python-dotenv = "*"
openai = "*"
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
bcrypt = "4.0.1"

